chatbot-api/requirement.txt
#fastapi -- Not required 
flask==3.1.1
flask_cors==6.0.0
uvicorn[standard]
chromadb
PyPDF2
python-docx
pandas
python-multipart
pydantic
streamlit
openai
requests
pymysql
python-dotenv